Adds save and open menu functionality
Adds a new class called FileMenuUtilities that saves and opens the application menu. This is achieved by using an ArrayList in the main JavaLauncher class that contains all of the ApplicationMenuItem objects. This arraylist can then be written to a file on the disk and retrieved on user demand. The ApplicationMenuItem class has also been modified to implement Serialization so the objects can be written to disk.